+++ service-worker.js
@@ -1,4 +1,4 @@
-const CACHE = 'click360-commercial-1-0-5-r3';
+const CACHE = 'click360-commercial-1-0-5-r4';
 const ASSETS = [
   './',
   './index.html',